Add option to exclude images by a javascript expression
Javascript that runs inside a jquery ".not(function() { ... })"
expression, where "this" is the a.colorbox element. The expression has
to return either "true" for removing an item from the set or "false" to
keep an item in the set. An example would bt to exlude certain classes:
"return $(this).hasClass('slick-cloned');"