Fix bug when ftw.colorbox is not installed but in the path.
The problem is thatt the pattern settings adapter was used even when
ftw.colorbox was not installed, causing exceptions when the
not-installed registry settings was read.

The solution is to use a browser layer and only provide the adapter
for this browser layer and thus when ftw.colorbox is installed.